together_back/pyproject.toml
[tool.poetry]
name = "together-back"
version = "0.1.0"
description = ""
authors = ["byungyong <quddyddlrk@naver.com>"]
readme = "README.md"
packages = [{include = "together_back"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
django-cors-headers = "^4.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
channels = "^4.0.0"
channels-redis = "^4.1.0"
python-decouple = "^3.8"
daphne = "^4.0.0"
mysqlclient = "^2.2.0"
django-allauth = "^0.56.1"
dj-rest-auth = "^5.0.1"


[tool.poetry.group.dev.dependencies]
daphne = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
